add figure scripts of NGWP JFAA paper and rearrange paperscripts folder
